Add a primitive search function.
Really, this is primitive. It search the convert HTML (!) by opening
all the mails and checking their contents... It would really be
better to dedicate this task to something like Lucene (but that
is Java...). Also, one shouldn’t search the HTML, but rather the
original mail sources, but sadly, there’s no way to map the email
source file names to the resulting HTML file names apart from
opening up **all** HTML files and checking for the message ID you
previously have to read from the email sourcefile which of course
has to be parsed previously. Improvements welcome.